Add generic Ubuntu system definition
Currently, NLOPT and GSL are enabled, but PETSc is not.  GSL's own CBLAS
library is used for BLAS to keep dependencies simple, and the compiler
optimization level is -O0 (this machine definition is intended for
development, not production).

Detection is performed using `lsb_release`, which may not be available
on all platforms (e.g. RHEL); more robust detection logic should be used
in the future.

This definition has been tested on Curran's laptop running Ubuntu 18.04
LTS (Bionic Beaver).